some changes

+++ inc/sorter.h
@@ -23,13 +23,11 @@ template <class T> class Sorter {
 
 public:
 
-  Sorter(size_t len);
+  Sorter(uint8_t len);
   ~Sorter();
 
   /// Sort an array
-  /// @param [in] arr the array to be sorted
-  /// @return number of iterations needed for sorting
-  unsigned long long sort(const T* arr);
+  uint64_t sort(const T* arr);
 
   /// Return the index array
   const uint8_t* getIndex(void);
@@ -37,12 +35,10 @@ public:
 
 private:
 
-  size_t   m_len;        ///< array length
+  uint8_t  m_len;        ///< array length
   uint8_t *m_index;      ///< index array
 
   /// Check if array is sorted
-  /// @param [in] arr the array to check
-  /// @return true if sorted, false otherwise
   bool isSorted(const T* arr);
 
   /// Shuffle the index array
@@ -50,14 +46,14 @@ private:
 
 };
 
-template <class T> Sorter<T>::Sorter(size_t len) {
+template <class T> Sorter<T>::Sorter(uint8_t len) {
   srand(time(NULL));
 
   m_len = len;
   m_index = (uint8_t*)malloc(m_len * sizeof(uint8_t));
 
   // Initialize in default order to catch pre-sorted arrays
-  for(size_t i = 0; i < m_len; i++) {
+  for (uint8_t i = 0; i < m_len; i++) {
     m_index[i] = i;
   }
 }
@@ -66,8 +62,8 @@ template <class T> Sorter<T>::~Sorter() {
   free(m_index);
 }
 
-template <class T> unsigned long long Sorter<T>::sort(const T* arr) {
-  unsigned long long i = 0;
+template <class T> uint64_t Sorter<T>::sort(const T* arr) {
+  uint64_t i = 0;
   while (!isSorted(arr)) {
     shuffleIndex();
     i++;
@@ -76,7 +72,7 @@ template <class T> unsigned long long Sorter<T>::sort(const T* arr) {
 }
 
 template <class T> bool Sorter<T>::isSorted(const T* arr) {
-  for (size_t i = 1; i < m_len; i++) {
+  for (uint8_t i = 1; i < m_len; i++) {
     if (arr[m_index[i]] < arr[m_index[i-1]]) {
       return false;
     }
@@ -85,9 +81,9 @@ template <class T> bool Sorter<T>::isSorted(const T* arr) {
 }
 
 template <class T> void Sorter<T>::shuffleIndex() {
-  for (size_t i = 0; i < m_len; i++) {
+  for (uint8_t i = 0; i < m_len; i++) {
     m_index[i] = (uint8_t)rand() % m_len;
-    for (size_t j = 0; j < i; j++) {
+    for (uint8_t j = 0; j < i; j++) {
       if (m_index[j] == m_index[i]) {
         i--;
         break;
